Organize organization code
