+++ src/Admin/AdminConsole/Models/ProviderEditModel.cs
@@ -2,7 +2,6 @@
 using Bit.Core.AdminConsole.Entities.Provider;
 using Bit.Core.AdminConsole.Enums.Provider;
 using Bit.Core.AdminConsole.Models.Data.Provider;
-using Bit.Core.Billing.Entities;
 using Bit.Core.Billing.Enums;
 using Bit.Core.Billing.Providers.Entities;
 using Bit.Core.Enums;
