+++ src/Api/Billing/Controllers/ProviderBillingController.cs
@@ -7,7 +7,6 @@ using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Providers.Models;
 using Bit.Core.Billing.Providers.Repositories;
 using Bit.Core.Billing.Providers.Services;
-using Bit.Core.Billing.Repositories;
 using Bit.Core.Billing.Services;
 using Bit.Core.Context;
 using Bit.Core.Models.BitStripe;
