+++ src/Api/Billing/Models/Responses/OrganizationMetadataResponse.cs
@@ -1,4 +1,5 @@
 ﻿using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Organizations.Models;
 
 namespace Bit.Api.Billing.Models.Responses;
