+++ src/Core/Billing/Organizations/Models/OrganizationSale.cs
@@ -1,8 +1,10 @@
 ﻿using Bit.Core.AdminConsole.Entities;
 using Bit.Core.Billing.Constants;
+using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Models.Sales;
 using Bit.Core.Models.Business;
 
-namespace Bit.Core.Billing.Models.Sales;
+namespace Bit.Core.Billing.Organizations.Models;
 
 #nullable enable
