+++ src/Core/Billing/Organizations/Repositories/IOrganizationInstallationRepository.cs
@@ -1,7 +1,7 @@
-﻿using Bit.Core.Billing.Entities;
+﻿using Bit.Core.Billing.Organizations.Entities;
 using Bit.Core.Repositories;
 
-namespace Bit.Core.Billing.Repositories;
+namespace Bit.Core.Billing.Organizations.Repositories;
 
 public interface IOrganizationInstallationRepository : IRepository<OrganizationInstallation, Guid>
 {
