+++ src/Core/Billing/Organizations/Services/IOrganizationBillingService.cs
@@ -1,8 +1,8 @@
 ﻿using Bit.Core.AdminConsole.Entities;
 using Bit.Core.Billing.Models;
-using Bit.Core.Billing.Models.Sales;
+using Bit.Core.Billing.Organizations.Models;
 
-namespace Bit.Core.Billing.Services;
+namespace Bit.Core.Billing.Organizations.Services;
 
 #nullable enable
