+++ src/Infrastructure.EntityFramework/Billing/Models/OrganizationInstallation.cs
@@ -4,7 +4,7 @@ using Bit.Infrastructure.EntityFramework.Platform;
 
 namespace Bit.Infrastructure.EntityFramework.Billing.Models;
 
-public class OrganizationInstallation : Core.Billing.Entities.OrganizationInstallation
+public class OrganizationInstallation : Core.Billing.Organizations.Entities.OrganizationInstallation
 {
     public virtual Installation Installation { get; set; }
     public virtual Organization Organization { get; set; }
@@ -14,6 +14,6 @@ public class OrganizationInstallationMapperProfile : Profile
 {
     public OrganizationInstallationMapperProfile()
     {
-        CreateMap<Core.Billing.Entities.OrganizationInstallation, OrganizationInstallation>().ReverseMap();
+        CreateMap<Core.Billing.Organizations.Entities.OrganizationInstallation, OrganizationInstallation>().ReverseMap();
     }
 }
