+++ src/Infrastructure.EntityFramework/Billing/Repositories/ProviderInvoiceItemRepository.cs
@@ -1,8 +1,6 @@
 ﻿using AutoMapper;
-using Bit.Core.Billing.Entities;
 using Bit.Core.Billing.Providers.Entities;
 using Bit.Core.Billing.Providers.Repositories;
-using Bit.Core.Billing.Repositories;
 using Bit.Infrastructure.EntityFramework.Repositories;
 using LinqToDB;
 using Microsoft.Extensions.DependencyInjection;
