+++ test/Api.Test/Billing/Controllers/OrganizationsControllerTests.cs
@@ -10,8 +10,8 @@ using Bit.Core.Auth.Enums;
 using Bit.Core.Auth.Models.Data;
 using Bit.Core.Auth.Repositories;
 using Bit.Core.Auth.Services;
+using Bit.Core.Billing.Organizations.Repositories;
 using Bit.Core.Billing.Pricing;
-using Bit.Core.Billing.Repositories;
 using Bit.Core.Billing.Services;
 using Bit.Core.Context;
 using Bit.Core.Entities;
